Fix surface creation and main path
