+++ src/input/mod.rs
@@ -1,0 +1,4 @@
+pub mod camera;
+pub mod keyboard;
+pub mod mouse;
+pub mod r#loop;
