+++ src/main.rs
@@ -1,0 +1,3 @@
+fn main() {
+    crate::platform::desktop::start();
+}
