+++ src/platform/desktop.rs
@@ -1,0 +1,7 @@
+#[cfg(not(target_arch = "wasm32"))]
+pub fn start() {
+    // desktop startup not implemented
+}
+
+#[cfg(target_arch = "wasm32")]
+pub fn start() {}
