+++ src/platform/mod.rs
@@ -1,0 +1,2 @@
+pub mod web;
+pub mod desktop;
