+++ src/render/init.rs
@@ -1,0 +1,47 @@
+use wasm_bindgen::JsValue;
+
+pub struct GpuContext {
+    pub surface: wgpu::Surface<'static>,
+    pub device: wgpu::Device,
+    pub queue: wgpu::Queue,
+    pub config: wgpu::SurfaceConfiguration,
+    pub depth_format: wgpu::TextureFormat,
+}
+
+pub async fn init(canvas: &web_sys::HtmlCanvasElement) -> Result<GpuContext, JsValue> {
+    let instance = wgpu::Instance::default();
+    let surface = unsafe { instance.create_surface(&canvas) }
+        .map_err(|e| JsValue::from_str(&e.to_string()))?;
+    let adapter = instance
+        .request_adapter(&wgpu::RequestAdapterOptions {
+            power_preference: wgpu::PowerPreference::HighPerformance,
+            compatible_surface: Some(&surface),
+            force_fallback_adapter: false,
+        })
+        .await
+        .ok_or("adapter")?;
+    let (device, queue) = adapter
+        .request_device(
+            &wgpu::DeviceDescriptor {
+                label: None,
+                required_features: wgpu::Features::empty(),
+                required_limits: adapter.limits(),
+            },
+            None,
+        )
+        .await
+        .map_err(|e| JsValue::from_str(&e.to_string()))?;
+    let caps = surface.get_capabilities(&adapter);
+    let config = wgpu::SurfaceConfiguration {
+        usage: wgpu::TextureUsages::RENDER_ATTACHMENT,
+        format: caps.formats[0],
+        width: canvas.width(),
+        height: canvas.height(),
+        present_mode: caps.present_modes[0],
+        desired_maximum_frame_latency: 2,
+        alpha_mode: caps.alpha_modes[0],
+        view_formats: vec![],
+    };
+    surface.configure(&device, &config);
+    Ok(GpuContext { surface, device, queue, config, depth_format: wgpu::TextureFormat::Depth32Float })
+}
